Rename + replace mapping keys
- Rename namespace to MappingKey (singular)
- Move lib/keys to lib/mapping-key
