+++ packages/codec/lib/mapping-key/encode/index.ts
